Sse refactor (#112)
* refactor SSE, gitignore testresources

* refactor SSE, gitignore testresources

* refactor SSE, gitignore testresources

* refactor SSE, gitignore testresources

* refactor SSE, gitignore testresources

* refactor SSE, gitignore testresources

* fix testresource folder not exist case

+++ src/test/java/com/blck/MusicReleaseTracker/HelperDB.java
@@ -26,12 +26,16 @@ import java.sql.*;
 
 public class HelperDB {
 
+    final static Path testResources = Paths.get("src", "test", "testresources");
     final static String testDBpath = "jdbc:sqlite:" + Paths.get("src", "test", "testresources", "testdb.db");
     final static Path DBfilePath = Paths.get("src", "test", "testresources", "testdb.db");
 
     public static void redoTestDB() {
         try {
-            Files.delete(DBfilePath);
+            if (!Files.exists(testResources))
+                Files.createDirectory(testResources);
+            if (Files.exists(DBfilePath))
+                Files.delete(DBfilePath);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
